refactor: merge community test code (#164)
